feat: implement full support for sized integer types and 64-bit literals
- Added Int8, Int16, Int32, Int64 and their UInt counterparts to the language.
- Refactored Int to be 64-bit to match documentation and stdlib expectations.
- Updated Lexer and Parser to support 64-bit integer literals using int64_t and std::stoull.
- Implemented implicit casting and type promotion in Binder for mixed-integer arithmetic and assignments.
- Enhanced IR generation with signed/unsigned division (SDiv/UDiv) and comparison predicates.
- Added support for BoundCast in LLVM and ZIR codegen to handle bit-width conversions.
- Verified changes with a new comprehensive test suite for integer types and overflow behavior.

+++ src/ast/const/const_int.hpp
@@ -3,10 +3,10 @@
 
 class ConstInt : public ExpressionNode {
 public:
-  int value_;
+  int64_t value_;
   std::string typeName_ = "i32";
   ConstInt() = default;
-  ConstInt(int value, std::string typeName = "i32")
+  ConstInt(int64_t value, std::string typeName = "i32")
       : value_(value), typeName_(typeName) {}
 
   void accept(Visitor &v) override { v.visit(*this); }
